Har lavet ændringer til klasser og DTO'er, så gjort DAO klar til at test om det hele kommer ind i databasen. Har ikke fået det til at køre endnu

+++ src/main/java/dat/DTO/ActorDTO.java
@@ -14,12 +14,13 @@ import java.time.LocalDate;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ActorDTO {
 
-    private long id;
+    private int id;
     @JsonProperty ("birth_date")
     private String birthDate;
     private String name;
     @JsonProperty("cast_id")
     private int actorId;
     private String character;
+
 
 }
